requirements.txt
opencv
pytorch
torchvision
pandas
numpy
skorch